@sanderegg review: refactor schemas
